Ported the main library to Windows
Finished porting the remaining files in the main library codebase to work
on Windows. Most of the edits were due to incompatibilities with the
in_addr_t structure.

Again, this code is yet to be tested.

+++ src/groundlift/obex.h
@@ -41,7 +41,7 @@ extern "C" {
  *
  * @param opcode Opcode to have its final bit set.
  */
-#define OBEX_SET_FINAL_BIT(opcode) ((opcode) | (1 << 7))
+#define OBEX_SET_FINAL_BIT(opcode) (uint8_t)((opcode) | (1 << 7))
 
 /**
  * Checks wether the opcode has the Final Bit Flag set.
